+++ src/App.vue
@@ -1,6 +1,6 @@
 <template>
 	<div class="flex justify-center pt-6">
-		<select v-model="selectedPreview" class="block w-full max-w-sm appearance-none rounded-md bg-white px-3 py-2 text-grey-900 shadow-sm outline-none ring-1 ring-inset ring-grey-300 placeholder:text-grey-400 focus:ring-2 focus:ring-purple-800">
+		<select v-model="selectedPreview" class="block w-full max-w-sm appearance-none rounded-md bg-white px-3 py-2 text-grey-900 shadow-sm outline-none ring-1 ring-inset ring-grey-300 placeholder:text-grey-400 focus:ring-2 focus:ring-purple-800 dark:bg-white/20 dark:text-grey-50 dark:focus:ring-purple-400">
 			<option v-for="option in previewOptions" :key="option" :value="option.value">
 				{{ option.label }}
 			</option>
@@ -13,12 +13,13 @@
 <script setup>
 import { ref } from "vue";
 
-const selectedPreview = ref("form-wrapper-preview");
+const selectedPreview = ref("dropdown-menu-preview");
 
 const previewOptions = [
 	{ label: "alert-message", value: "alert-message-preview" },
 	{ label: "button-group", value: "button-group-preview" },
 	{ label: "donut-chart", value: "donut-chart-preview" },
+	{ label: "dropdown-menu", value: "dropdown-menu-preview" },
 	{ label: "form-checkbox", value: "form-checkbox-preview" },
 	{ label: "form-input", value: "form-input-preview" },
 	{ label: "form-textarea", value: "form-textarea-preview" },
@@ -94,7 +95,19 @@ select {
 	padding-right: 2.5em;
 }
 
+.dark select {
+	background-image: url("data:image/svg+xml,%3csvg xmlns='http://www.w3.org/2000/svg' fill='none' viewBox='0 0 20 20'%3e%3cpath stroke='%23ffffff40' stroke-linecap='round' stroke-linejoin='round' stroke-width='1.5' d='M6 8l4 4 4-4'/%3e%3c/svg%3e");
+	background-position: right 0.5rem center;
+	background-repeat: no-repeat;
+	background-size: 1.5em 1.5em;
+	padding-right: 2.5em;
+}
+
 select:focus {
 	background-image: url("data:image/svg+xml,%3csvg xmlns='http://www.w3.org/2000/svg' fill='none' viewBox='0 0 20 20'%3e%3cpath stroke='%236b21a8' stroke-linecap='round' stroke-linejoin='round' stroke-width='1.5' d='M6 8l4 4 4-4'/%3e%3c/svg%3e");
+}
+
+.dark select:focus {
+	background-image: url("data:image/svg+xml,%3csvg xmlns='http://www.w3.org/2000/svg' fill='none' viewBox='0 0 20 20'%3e%3cpath stroke='%23c084fc' stroke-linecap='round' stroke-linejoin='round' stroke-width='1.5' d='M6 8l4 4 4-4'/%3e%3c/svg%3e");
 }
 </style>