+++ src/components/interaction/dropdown-menu/fragments/dropdown-menu-button.vue
@@ -1,0 +1,5 @@
+<template>
+	<ui-button class="w-full border-s-2 border-transparent px-4 py-2 hocus:border-current hocus:bg-grey-50 hocus:text-purple-800 dark:hocus:bg-white/10 dark:hocus:text-purple-300">
+		<slot />
+	</ui-button>
+</template>