+++ src/components/interaction/dropdown-menu/fragments/dropdown-menu-divider.vue
@@ -1,0 +1,3 @@
+<template>
+	<div class="my-2 border-b border-grey-200 dark:border-white/20" />
+</template>