+++ src/components/interaction/dropdown-menu/fragments/dropdown-menu-link.vue
@@ -1,0 +1,5 @@
+<template>
+	<link-tag class="w-full border-s-2 border-transparent px-4 py-2 hocus:border-current hocus:bg-grey-50 dark:hocus:bg-white/10">
+		<slot />
+	</link-tag>
+</template>