+++ src/preview/preview-section.vue
@@ -1,5 +1,5 @@
 <template>
-	<div class="mb-6 border-b border-grey-200 pb-6 dark:border-grey-700">
+	<div class="mb-6 border-b border-grey-200 pb-6 dark:border-white/20">
 		<h2 class="mb-4 text-lg font-bold text-grey-950 dark:text-grey-50">
 			<slot name="title" />
 		</h2>